Добавил новый компонент nearby-block
