+++ project/src/pages/property/nearby/nearby-rooms.tsx
@@ -1,23 +1,9 @@
 import { memo } from 'react';
 import PlaceCard from '../../../components/place-card/place-card';
-import Spinner from '../../../pages/loading-screen/spinner';
 import { useAppSelector } from '../../../hooks';
 
 function NearbyRooms(): JSX.Element {
   const serverNearbyOffers = useAppSelector((state) => state.serverNearbyOffers);
-  const isNearbyOffersDataLoading = useAppSelector((state) => state.isNearbyOffersDataLoading);
-
-  if (isNearbyOffersDataLoading) {
-    const spinnerSize = {
-      width: 1000,
-      height: 424,
-    };
-    return (
-      <section className="near-places places">
-        <Spinner spinnerSize={[spinnerSize.height, spinnerSize.width]}/>
-      </section>
-    );
-  }
 
   return (
     <section className="near-places places">
